Move Content-related models into a subdirectory
This change moves the following classes from `models` into a
`models/content` subdirectory, with a corresponding `Content` module:

* `ContentItem`
* `Group`
* `Link`
