+++ app/decorators/content/items_decorator.rb
@@ -1,0 +1,14 @@
+module Content
+  class ItemsDecorator < Draper::CollectionDecorator
+    delegate(
+      :current_page,
+      :total_pages,
+      :limit_value,
+      :entry_name,
+      :total_count,
+      :offset_value,
+      :last_page?,
+      :next_item,
+    )
+  end
+end
