--- app/decorators/content_items_decorator.rb
@@ -1,12 +1,0 @@
-class ContentItemsDecorator < Draper::CollectionDecorator
-  delegate(
-    :current_page,
-    :total_pages,
-    :limit_value,
-    :entry_name,
-    :total_count,
-    :offset_value,
-    :last_page?,
-    :next_item,
-  )
-end
