+++ app/domain/content/importers/all_google_analytics_metrics.rb
@@ -7,7 +7,7 @@ module Content
     end
 
     def run
-      ContentItem.find_in_batches(batch_size: batch_size) do |content_items|
+      Item.find_in_batches(batch_size: batch_size) do |content_items|
         ImportPageviewsJob.perform_later(content_items)
       end
     end
