+++ app/domain/content/importers/pageviews.rb
@@ -13,7 +13,7 @@ module Content
 
       results = @google_analytics_service.page_views(base_paths)
       results.each do |result|
-        content_item = ContentItem.find_by(base_path: result[:base_path])
+        content_item = Item.find_by(base_path: result[:base_path])
         content_item.update!(result.slice(:one_month_page_views, :six_months_page_views))
       end
     end
