+++ app/domain/content/rules_filter.rb
@@ -9,7 +9,7 @@ module Content
     def apply(scope)
       groups = rules.group_by(&:link_type)
 
-      groups.inject(ContentItem.none) do |filtered_scope, (link_type, group_rules)|
+      groups.inject(Item.none) do |filtered_scope, (link_type, group_rules)|
         target_ids = group_rules.map(&:target_content_id)
         link_filter = LinkFilter.new(link_type: link_type, target_ids: target_ids)
