+++ spec/domain/metric_builder_spec.rb
@@ -13,7 +13,7 @@ module Performance
         create :content_item, number_of_pdfs: 1
         create :content_item, one_month_page_views: 2
 
-        result = subject.run_collection(ContentItem.all)
+        result = subject.run_collection(Content::Item.all)
         expected_result = {
           total_pages: { value: 2 },
           zero_page_views: { value: 1 },
