+++ spec/factories.rb
@@ -2,7 +2,7 @@ require_relative "../app/models/audits/question"
 require_relative "./factories/link_factory"
 
 FactoryGirl.define do
-  factory :content_item do
+  factory :content_item, class: Content::Item do
     transient do
       organisations nil
       primary_publishing_organisation nil
@@ -33,7 +33,7 @@ FactoryGirl.define do
     end
   end
 
-  factory :link do
+  factory :link, class: Content::Link do
     sequence(:source_content_id) { |i| "source-#{i}" }
     sequence(:target_content_id) { |i| "target-#{i}" }
     link_type "organisations"
